api/requirements.txt
fastapi>=0.109
uvicorn[standard]>=0.27
pandas>=2.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
